example/requirements.txt
openai
requests
PyPDF2
sentence-transformers
faiss-cpu
numpy
python-dotenv